Fixed memory UD bug in Clone for Cif.
The derived impl of Clone for Cif doesn't work because while it clones
the Type and TypeArray in the Cif, it doesn't make the copy of the
ffi_cif in the clone refer to the cloned Type and TypeArray. Now I've
written an explicit impl should do the right thing.