fix soundness of `high` interface, around `ClosureN::code_ptr`
The `code_ptr` methods were unsound, as bare `fn` pointers are Copy,
and so returning references with appropriately non-'static lifetimes
to them will not actually work to prevent their usage outside of those
lifetimes. Instead, we now use an ABI compatible wrapper type that
carries that lifetime in its definition, preventing that misuse.