Add tests for ClosureOnce.
We check that we can call a ClosureOnce once successfully,
but it’s an error to call it twice. Normally if this error
occurs, the program exits (code 2), because it isn’t
safe to panic when calling a callback. However, in
cfg(test) this error results in a panic so that we can
test the error using #[should_panic].