Fix return type in call example code
In the example code provided as part of the documention of `call`,
the return type is not explicitly specified, and will be inferred
by the Rust compiler as `i32`.  However, the actual C routine
being invoked has a `u64` return type.  This can cause the test
to fail on big-endian platforms.

Fixed by explicitly specifying the `u64` return type.