Clean up the ancient CI config
This refactors CI such that we only run on jobs that don't use QEMU, and
such that we no longer depend on ancient versions of unmaintained
Actions projects.

The removal of QEMU jobs is due to these jobs failing randomly rather
often, and thus aren't particularly useful.

The Windows jobs are also changed to not install a whole bunch of extra
targets, as the CI setup was such that these targets weren't actually
used to begin with.