Select build process by `CARGO_CFG_TARGET_ENV`
`target_env` represent the host building the code for build scripts, and
does not necessarily reflect the target platform. This change uses
`CARGO_CFG_TARGET_ENV` to select the correct build process.

Currently, it is seemingly not possible to cross-compile libffi-sys for
MSVC, but it might be possible in the future.