Change type of `FFI_TYPE_` constants to `u16`
Type tags are stored in the `type_` field of `ffi_type`s, which is a
`c_ushort`. `c_ushort` is an unsigned integer with at least 16 bits. On
most modern architectures, `c_ushort` is equivalent to `u16`.

This commit changes the types of the constants used for the `type_`
field to `u16` to avoid having to cast the constants when using them.